chore: rewrite BetterTTV
