+++ layers/emote-integrations/layers/integrations/nuxt.config.ts
@@ -24,16 +24,24 @@ export default defineNuxtConfig({
   extends: allLowercaseEmoteSources,
   imports: {
     imports: [
-      ...allEmoteSources_.map((source) => ({
-        from: `#t_${toLowerCase(source)}`,
-        name: `T${source}`,
-        type: true,
-      })),
+      ...allEmoteSources_
+        .filter((source) => source !== "BetterTTV")
+        .map((source) => ({
+          from: `#t_${toLowerCase(source)}`,
+          name: `T${source}`,
+          type: true,
+        })),
       ...allEmoteSources_.map((source) => ({
         from: `#integrations_${toLowerCase(source)}`,
         name: `${source}Api`,
         type: true,
       })),
+      {
+        from: "#t_betterttv",
+        name: "*",
+        as: "TBetterTTV",
+        type: true,
+      },
     ],
   },
 });
